+++ packages/zefyr/example/lib/src/full_page.dart
@@ -94,6 +94,7 @@ class _FullPageEditorScreenState extends State<FullPageEditorScreen> {
             focusNode: _focusNode,
             mode: _editing ? ZefyrMode.edit : ZefyrMode.select,
             imageDelegate: CustomImageDelegate(),
+            keyboardAppearance: Brightness.dark,
           ),
         ),
       ),