+++ packages/zefyr/lib/src/widgets/input.dart
@@ -24,15 +24,16 @@ class InputConnectionController implements TextInputClient {
 
   /// Opens or closes input connection based on the current state of
   /// [focusNode] and [value].
-  void openOrCloseConnection(FocusNode focusNode, TextEditingValue value) {
+  void openOrCloseConnection(FocusNode focusNode, TextEditingValue value,
+      Brightness keyboardAppearance) {
     if (focusNode.hasFocus && focusNode.consumeKeyboardToken()) {
-      openConnection(value);
+      openConnection(value, keyboardAppearance);
     } else if (!focusNode.hasFocus) {
       closeConnection();
     }
   }
 
-  void openConnection(TextEditingValue value) {
+  void openConnection(TextEditingValue value, Brightness keyboardAppearance) {
     if (!hasConnection) {
       _lastKnownRemoteTextEditingValue = value;
       _textInputConnection = TextInput.attach(
@@ -42,6 +43,7 @@ class InputConnectionController implements TextInputClient {
           obscureText: false,
           autocorrect: true,
           inputAction: TextInputAction.newline,
+          keyboardAppearance: keyboardAppearance,
           textCapitalization: TextCapitalization.sentences,
         ),
       )..setEditingState(value);