+++ packages/zefyr/example/lib/main.dart
@@ -2,6 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 import 'package:flutter/material.dart';
+
 import 'src/form.dart';
 import 'src/full_page.dart';
 import 'src/view.dart';
@@ -16,7 +17,6 @@ class ZefyrApp extends StatelessWidget {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
       title: 'Zefyr Editor',
-      theme: ThemeData(primarySwatch: Colors.cyan),
       home: HomePage(),
       routes: {
         "/fullPage": buildFullPage,
@@ -44,32 +44,21 @@ class HomePage extends StatelessWidget {
   Widget build(BuildContext context) {
     final nav = Navigator.of(context);
     return Scaffold(
-      appBar: AppBar(
-        elevation: 1.0,
-        backgroundColor: Colors.grey.shade200,
-        brightness: Brightness.light,
-        title: ZefyrLogo(),
-      ),
+      appBar: AppBar(title: ZefyrLogo()),
       body: Column(
         children: <Widget>[
           Expanded(child: Container()),
-          FlatButton(
+          RaisedButton(
             onPressed: () => nav.pushNamed('/fullPage'),
             child: Text('Full page editor'),
-            color: Colors.lightBlue,
-            textColor: Colors.white,
           ),
-          FlatButton(
+          RaisedButton(
             onPressed: () => nav.pushNamed('/form'),
             child: Text('Embedded in a form'),
-            color: Colors.lightBlue,
-            textColor: Colors.white,
           ),
-          FlatButton(
+          RaisedButton(
             onPressed: () => nav.pushNamed('/view'),
             child: Text('Read-only embeddable view'),
-            color: Colors.lightBlue,
-            textColor: Colors.white,
           ),
           Expanded(child: Container()),
         ],